Update hangman.py
Hi Derek, Levi here. When Bear and myself were at your presentation, thanks again for that, we came across two issues in your hangman game so we decided to dig in and try and fix it. We just replaced the " ' "/" é " with " "/" e ". We felt that getting rid of the offending words seemed to be the best course of action for the problem but aren't knowledgeable enough yet lol. We also want to try and get the guessed word to print if the guesses run out but that's going to have to wait till tomorrow cause he has homework and my brain already feels fried lol.